Extracted song-requesting to a different class
